requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.0.1
SpeechRecognition==3.10.1
pydub==0.25.1
google-speech==1.2.0
requests==2.31.0